+++ app/src/main/java/com/tomerpacific/moviepresenter/domain/repository/MovieRepositoryImpl.kt
@@ -3,7 +3,7 @@ package com.tomerpacific.moviepresenter.domain.repository
 import android.graphics.BitmapFactory
 import com.tomerpacific.moviepresenter.BuildConfig
 import com.tomerpacific.moviepresenter.domain.model.MovieModel
-import com.tomerpacific.moviepresenter.model.TMDBResponse
+import com.tomerpacific.moviepresenter.domain.model.TMDBResponse
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.coroutineScope
 import kotlinx.coroutines.launch
